+++ HarfangLab/harfanglab/models.py
@@ -6,7 +6,7 @@ Data models of the HarfangLab module
 from typing import Any
 
 # third parties
-from pydantic.v1 import BaseModel
+from pydantic import BaseModel
 
 
 class JobTarget(BaseModel):